Changes to the build process in order to run tests.
The ./test directory contains the redex tests published by facebook.
Unfortunately, the build files (Makefile.am) in this directory are out-of-date
compared to the current state of the code proper, with the result that tests
do not compile/link correctly.

This is probably going to be fixed by facebook redex at some point, but until then
there are some changes in this commit that will allow most of the tests to build
(with some effort).

Eventually, a testing solution for clue-redex will probably merge with these
tests.